requirements.txt
requests
pandas
boto3
python-dotenv
pyarrow